new mangainfoview
