+++ Manga/Features/Manga/MangaInfoView.swift
@@ -14,29 +14,35 @@ struct MangaInfoView: View {
     
     var body: some View {
         NavigationStack {
-            ScrollView {
-                ZStack {
-                    backgroundImage
-                    frontContent
-                }
+            ZStack {
+                backgroundImage
+                frontContent
             }
-            .onAppear() {
-                
-            }
-            .navigationTitle(mangaVM.manga?.data?.attributes?.title?.en ?? "")
+            .navigationTitle(manga.attributes?.title?.en ?? "")
         }
     }
     
-    var backgroundImage: some View {
+    var frontContent: some View {
+        ScrollView {
+            VStack {
+                HStack {
+                    frontImage
+                    titleStack
+                    Spacer()
+                }
+            }
+        }
+    }
+    
+    var frontImage: some View {
         AsyncImage(
             url: URL(string: mangaVM.generateCoverURL(manga: manga)),
             content: { image in
                 image.resizable()
-                    .scaledToFill()
-                    .frame(height: 350, alignment: .top)
-                    .opacity(0.5)
-                    .clipped()
-                    .blur(radius: 3)
+                    .aspectRatio(contentMode: .fit)
+                    .frame(maxWidth: 300, maxHeight: 450)
+                    .padding(.leading, 35)
+                    .padding(.top, 20)
             },
             placeholder: {
                 ProgressView()
@@ -44,22 +50,34 @@ struct MangaInfoView: View {
         )
     }
     
-    var frontContent: some View {
+    var titleStack: some View {
         VStack {
-            Text("\(manga.id)")
             Text("\(manga.attributes?.title?.en ?? "")")
+                .font(.largeTitle)
+            Text("\(OriginalLanguage(rawValue: manga.attributes?.originalLanguage ?? "")?.getFlag() ?? "")")
+                
+            Spacer()
+        }
+        .padding(.top, 20)
+    }
+    
+    var backgroundImage: some View {
+        VStack {
             AsyncImage(
                 url: URL(string: mangaVM.generateCoverURL(manga: manga)),
                 content: { image in
                     image.resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .frame(maxWidth: 150, maxHeight: 200)
-                        .padding(.trailing, 10)
+                        .scaledToFill()
+                        .frame(height: 350, alignment: .top)
+                        .opacity(0.5)
+                        .clipped()
+                        .blur(radius: 3)
                 },
                 placeholder: {
                     ProgressView()
                 }
             )
+            Spacer()
         }
     }
 }
